refactor(crawler): move sitemap XML parsing to infrastructure (#442)
parse_sitemap is a quick-xml streaming parser — an infrastructure concern
that was living in the application layer. Move it to
infrastructure/crawler/sitemap_parser.rs alongside SitemapParser (which
already imports quick-xml, Url and CrawlError).

Re-exported from infrastructure/crawler/mod.rs and from discovery.rs so the
crawler facade, the bench (webfang_core::application::crawler::parse_sitemap)
and the fuzz target (...::discovery::parse_sitemap) resolve unchanged. The
existing unit tests stay in discovery.rs and now exercise the re-export;
all pass unmodified. Pure relocation, no behavior change.
